add function to check valid customer + create synthetic data 100 records

+++ process.py
@@ -47,10 +47,14 @@ def event_processor(evt: dict):
         # use the CHECK_DATA_QUALITY event processor dedicated function
         logger.info(f"Use the check data quality event processor")
         check_data_quality_contracts_event_processor(evt)
-    elif evt_type == "CHECK_COMMON_DEMO_CUSTOMERS":
+    elif evt_type == "CHECK_COMMON_CUSTOMERS":
         # use the CHECK_COMMON__DEMO_CUSTOMERS event processor dedicated function
-        logger.info(f"Use the check common customers demo event processor")
-        check_common_customers_demo_event_processor(evt)
+        logger.info(f"Use the check common customers  event processor")
+        check_common_customers_event_processor(evt)
+    elif evt_type == "CHECK_VALID_CUSTOMER":
+        # use the CHECK_VALID_CUSTOMER event processor dedicated function
+        logger.info(f"Use the check valid customer  event processor")
+        check_valid_customer_event_processor(evt)
     else:
         # use the GENERIC event processor function, that basicaly does nothing
         logger.info(f"Unhandled message type, use the generic event processor")
@@ -217,7 +221,6 @@ def fuse_event_processor(evt: dict):
                     None
                 )
                 participant=client_id
-
                 for index, row in encrypted_data.iterrows():
                     commutative_encrypt=tee_commutative_encrypt(row['customer_email'],participant,public_keys,n)
                     query="UPDATE customers_list_"+str(i)+" SET commutative_id = '"+str(commutative_encrypt)+"' WHERE customers_list_"+str(i)+".customer_email='"+row['customer_email']+"'"
@@ -247,7 +250,7 @@ def check_data_quality_contracts_event_processor(evt: dict):
     except Exception as e:
         logger.error(e)
 
-def check_common_customers_demo_event_processor(evt: dict):
+def check_common_customers_event_processor(evt: dict):
     logger.info(f"---------------------------------------------------------")
     logger.info(f"|                    START PROCESSING                   |")
     logger.info(f"|                                                       |")
@@ -278,6 +281,63 @@ def check_common_customers_demo_event_processor(evt: dict):
             logger.info(f"| 3. Send output                                        |")
             output_json={}
             output_json["common_customers"]={"by_email":common_customers_by_email}
+            with open(default_settings.data_user_output_location+'/report.json', 'w', newline='') as file:
+                    file.write(json.dumps(output_json, indent=4))
+            logger.info(f"|                                                       |")
+            execution_time=(time.time() - start_time)
+            logger.info(f"|    Execution time:  {execution_time} secs           |")
+            logger.info(f"|                                                       |")
+            logger.info(f"--------------------------------------------------------")
+
+        else:
+            logger.error(f"No table exist in memory, please initialise the fusion")
+    except Exception as e:
+        logger.error(e) 
+    
+def check_valid_customer_event_processor(evt: dict):
+    logger.info(f"---------------------------------------------------------")
+    logger.info(f"|                    START PROCESSING                   |")
+    logger.info(f"|                                                       |")
+    start_time = time.time()
+    logger.info(f"|    Start time:  {start_time} secs               |")
+    logger.info(f"|                                                       |")
+    audit_log(f"Start processing event: {evt.get('type', '')}.",LogLevel.INFO)
+    try:
+        logger.info(f"| 2. Load keys and fuse parameter                       |")
+        logger.info(f"|                                                       |")
+        with open(default_settings.data_connector_config_location+'/shared_modulus.json') as f:
+            n = json.load(f)["n"] 
+        with open(default_settings.data_connector_config_location+'/public_keys.json') as f:
+            public_keys = json.load(f)
+        
+        email= evt.get("email", "")
+        #TODO need to load participant (parameter sender) dynamically
+        participant="66e1a419eb0cbee048a2bce3"
+        commutative_email=tee_commutative_encrypt(email,participant,public_keys,n)
+                    
+        logger.info(f"| 1. Check valid customers                              |")
+        logger.info(f"|                                                       |")
+        #Connect in memory duckdb (encrypted memory on confidential computing)
+        con = duckdb.connect(database=":memory:")
+        con.sql("IMPORT DATABASE '"+default_settings.data_connector_config_location+"'") 
+        #check if tables exist in memory
+        existing_tables=con.sql("SHOW ALL TABLES; ")
+        if len(existing_tables)>0:
+            #check common customers by email in the database in memory
+            #Common customers by email
+            #Create duckdb query
+            query="SELECT COUNT(*) as total FROM customers_list_0,customers_list_1 WHERE (customers_list_0.commutative_id='"+str(commutative_email)+"' AND customers_list_1.commutative_id='"+str(commutative_email)+"')"
+            df = con.sql(query).df()
+            valid_customers_found="false"
+            if int(df["total"].to_string(index=False))>0:
+                valid_customers_found="true"
+
+            #Write outputs for data user
+            #For now the output is written in an encrypted drive only accessible for data user
+            #TODO Connector for data users (write) have to be created
+            logger.info(f"| 3. Send output                                        |")
+            output_json={}
+            output_json["valid_customer"]=valid_customers_found
             with open(default_settings.data_user_output_location+'/report.json', 'w', newline='') as file:
                     file.write(json.dumps(output_json, indent=4))
             logger.info(f"|                                                       |")
